fix cosine bug

+++ ait/components/debug/compare/msquickcmp/pta_acl_cmp/cmp_algorithm.py
@@ -20,8 +20,8 @@ NAN = 'NaN'
 
 
 def cosine_similarity(pta_data: np.ndarray, acl_data: np.ndarray):
-    pta_data = pta_data.reshape(-1)
-    acl_data = acl_data.reshape(-1)
+    pta_data = pta_data.reshape(-1).astype("float32")
+    acl_data = acl_data.reshape(-1).astype("float32")
 
     acl_data_norm = np.linalg.norm(acl_data, axis=-1, keepdims=True)
     pta_data_norm = np.linalg.norm(pta_data, axis=-1, keepdims=True)
